requirements.txt
fastai==1.0.51
opencv-python>=3.3.0.10
pillow
tornado~=5.1.0
